REFACTOR : Code Refactor according to new schema

+++ MakeProfits.Backend/MakeProfits.Backend/Models/Advisor.cs
@@ -2,6 +2,6 @@
 {
     public class Advisor : AbstractUser
     {
-        public int Rating { get; set; }
+        public decimal Rating { get; set; }
     }
 }
